Se crean las clases restantes y se hace prueba de cita

+++ Modules/grupo02/src/main/java/grupo02/Person.java
@@ -6,14 +6,14 @@ import lombok.Data;
 public class Person {
     private String name;
     private String lastName;
-    private int phone;
+    private long phone;
     private String email;
 
     public Person(){
 
     }
 
-    public Person(String name, String lastName, int phone, String email){
+    public Person(String name, String lastName, long phone, String email){
         this.name = name;
         this.lastName = lastName;
         this.phone = phone;
